Revert "update the comment"
