+++ doc/en/user/source/styling/sld/extensions/index.rst
@@ -19,4 +19,3 @@ Although not portable to other applications, these extensions make styling more 
    randomized
    composite-blend/index
    z-order/index
-   rendering-selection
