+++ src/extension/app-schema/app-schema-test/pom.xml
@@ -300,8 +300,8 @@
                 <exclude>**/IdFunctionWfsWithJoiningTest.java</exclude>
                 <exclude>**/XPathPredicateTest.java</exclude>
                 <exclude>${test.exclude.pattern}</exclude>
-                <exclude>**/GetFeatureNumberMatchedGMLTest.java</exclude>
-                <exclude>**/GetFeatureNumberMatchedGeoJSONTest.java</exclude>
+                <exclude>**/GetFeaturesNumberMatchedTest.java</exclude>
+                <exclude>**/GetFeatureNumberMatchedFeatureChainingTest.java</exclude>
                 <exclude>**/FeatureChainingWfsTest.java</exclude>
                 <exclude>**/NoIdExpressionTest.java</exclude>
               </excludes>
