+++ src/wms/src/main/java/org/geoserver/wms/legendgraphic/JSONLegendGraphicBuilder.java
@@ -269,7 +269,6 @@ public class JSONLegendGraphicBuilder extends LegendGraphicBuilder {
             FeatureType layer = legend.getFeatureType();
 
             Style gt2Style = legend.getStyle();
-            gt2Style = applyRenderingSelection(gt2Style);
             if (gt2Style == null) {
                 throw new NullPointerException("request.getStyle()");
             }
