+++ src/wms/src/test/java/org/geoserver/wms/legendgraphic/BaseLegendTest.java
@@ -47,8 +47,6 @@ public class BaseLegendTest<T extends LegendGraphicBuilder> extends WMSTestSuppo
         testData.addStyle("rainfall_ramp", MockData.class, catalog);
         testData.addStyle("rainfall_classes", MockData.class, catalog);
         testData.addStyle("rainfall_classes_nolabels", MockData.class, catalog);
-        testData.addStyle("styleWithLegendSelection", this.getClass(), catalog);
-        testData.addStyle("styleWithLegendSelectionOnSymbolizer", this.getClass(), catalog);
         // add raster layer for rendering transform test
         testData.addRasterLayer(
                 new QName("http://www.opengis.net/wcs/1.1.1", "DEM", "wcs"),
