+++ src/wms/src/test/java/org/geoserver/wms/legendgraphic/BufferedImageLegendGraphicOutputFormatTest.java
@@ -14,9 +14,7 @@ import static org.junit.Assert.fail;
 import java.awt.Color;
 import java.awt.image.BufferedImage;
 import java.awt.image.RenderedImage;
-import java.io.File;
 import java.io.IOException;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -45,7 +43,6 @@ import org.geotools.feature.type.AttributeDescriptorImpl;
 import org.geotools.feature.type.AttributeTypeImpl;
 import org.geotools.feature.type.GeometryDescriptorImpl;
 import org.geotools.feature.type.GeometryTypeImpl;
-import org.geotools.image.test.ImageAssert;
 import org.geotools.image.util.ImageUtilities;
 import org.geotools.referencing.CRS;
 import org.geotools.renderer.lite.RendererUtilities;
@@ -1455,52 +1452,6 @@ public class BufferedImageLegendGraphicOutputFormatTest
         this.legendProducer.resizeForDPI(req, sldStype);
         assertEquals(66, this.legendProducer.w);
         assertEquals(66, this.legendProducer.h);
-    }
-
-    @org.junit.Test
-    public void testLegendRenderingSelectionInRule() throws Exception {
-        // test <VendorOption name=renderingLegend>false</VendorOption> set for rule
-        // works for a getLegendGraphicRequest
-        // only the yellow square should be displayed
-        GetLegendGraphicRequest req = new GetLegendGraphicRequest(null);
-        req.setWidth(20);
-        req.setHeight(20);
-
-        FeatureTypeInfo ftInfo =
-                getCatalog()
-                        .getFeatureTypeByName(
-                                MockData.MPOLYGONS.getNamespaceURI(),
-                                MockData.MPOLYGONS.getLocalPart());
-
-        req.setLayer(ftInfo.getFeatureType());
-        req.setStyle(getCatalog().getStyleByName("styleWithLegendSelection").getStyle());
-
-        BufferedImage image = this.legendProducer.buildLegendGraphic(req);
-        URL result = getClass().getResource("./results/renderingSelectionOnRule.png");
-        ImageAssert.assertEquals(new File(result.toURI()), image, 50);
-    }
-
-    @org.junit.Test
-    public void testLegendRenderingSelectionInSymbolizer() throws Exception {
-        // test the <VendorOption name=renderingLegend>false</VendorOption> in a Symbolizer
-        // only the a blue square should be rendered
-        GetLegendGraphicRequest req = new GetLegendGraphicRequest(null);
-        req.setWidth(20);
-        req.setHeight(20);
-
-        FeatureTypeInfo ftInfo =
-                getCatalog()
-                        .getFeatureTypeByName(
-                                MockData.MPOLYGONS.getNamespaceURI(),
-                                MockData.MPOLYGONS.getLocalPart());
-
-        req.setLayer(ftInfo.getFeatureType());
-        req.setStyle(
-                getCatalog().getStyleByName("styleWithLegendSelectionOnSymbolizer").getStyle());
-
-        BufferedImage image = this.legendProducer.buildLegendGraphic(req);
-        URL result = getClass().getResource("./results/renderingSelectionOnSymbolizer.png");
-        ImageAssert.assertEquals(new File(result.toURI()), image, 50);
     }
 
     /** */
