+++ src/wms/src/test/java/org/geoserver/wms/legendgraphic/JSONLegendGraphicOutputFormatTest.java
@@ -1662,66 +1662,6 @@ public class JSONLegendGraphicOutputFormatTest extends BaseLegendTest<JSONLegend
 
         assertEquals("intervals", colormap.get(JSONLegendGraphicBuilder.COLORMAP_TYPE));
     }
-
-    @org.junit.Test
-    public void testLegendSelectionInRule() throws Exception {
-        // test that the <VendorOption name=renderingLegend>false</VendorOption> works with the Json
-        // output
-        GetLegendGraphicRequest req = getRequest();
-        req.setWidth(20);
-        req.setHeight(20);
-
-        FeatureTypeInfo ftInfo =
-                getCatalog()
-                        .getFeatureTypeByName(
-                                MockData.MPOLYGONS.getNamespaceURI(),
-                                MockData.MPOLYGONS.getLocalPart());
-
-        req.setLayer(ftInfo.getFeatureType());
-        req.setStyle(getCatalog().getStyleByName("styleWithLegendSelection").getStyle());
-        req.setFormat(JSONFormat);
-        JSONObject jsonLegend = this.legendProducer.buildLegendGraphic(req);
-        JSONArray legend = jsonLegend.getJSONArray(JSONLegendGraphicBuilder.LEGEND);
-        JSONArray rules = legend.getJSONObject(0).getJSONArray("rules");
-        // only one rule should be included in the output the other one has the renderingLegend
-        // option set to false.
-        assertEquals(1, rules.size());
-        String ruleName = rules.getJSONObject(0).getString("name");
-        assertEquals("nationalpark", ruleName);
-    }
-
-    @org.junit.Test
-    public void testLegendSelectionInSymbolizer() throws Exception {
-        // test that the <VendorOption name=renderingLegend>false</VendorOption> option set in
-        // symbolizer works
-        // avoiding the inclusion in the final output of one of the two symbolizers
-        GetLegendGraphicRequest req = getRequest();
-        req.setWidth(20);
-        req.setHeight(20);
-
-        FeatureTypeInfo ftInfo =
-                getCatalog()
-                        .getFeatureTypeByName(
-                                MockData.MPOLYGONS.getNamespaceURI(),
-                                MockData.MPOLYGONS.getLocalPart());
-
-        req.setLayer(ftInfo.getFeatureType());
-        req.setStyle(
-                getCatalog().getStyleByName("styleWithLegendSelectionOnSymbolizer").getStyle());
-        req.setFormat(JSONFormat);
-        JSONObject jsonLegend = this.legendProducer.buildLegendGraphic(req);
-        JSONArray rules =
-                jsonLegend
-                        .getJSONArray(JSONLegendGraphicBuilder.LEGEND)
-                        .getJSONObject(0)
-                        .getJSONArray("rules");
-        assertEquals(1, rules.size());
-        JSONArray symbolizers = rules.getJSONObject(0).getJSONArray("symbolizers");
-        // only one symbolizer of the two defined should be present in the final output
-        // the other one has the renderingLabel option set to false
-        assertEquals(1, symbolizers.size());
-    }
-
     /** @param result */
     private void assertNotEmpty(JSONObject result) {
         assertNotNull(result);
