--- src/wms/src/test/resources/org/geoserver/wms/legendgraphic/styleWithLegendSelection.sld
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sld:StyledLayerDescriptor
-xmlns="http://www.opengis.net/sld"
-xmlns:sld="http://www.opengis.net/sld"
-xmlns:ogc="http://www.opengis.net/ogc"
-xmlns:gml="http://www.opengis.net/gml"
-xmlns:xlink="http://www.w3.org/1999/xlink" version="1.0.0">
-  
-  <sld:UserLayer>
-    <sld:LayerFeatureConstraints>
-      <sld:FeatureTypeConstraint/>
-    </sld:LayerFeatureConstraints>
-    <sld:UserStyle>
-      <sld:Name>tl 2010 08013 arealm</sld:Name>
-      <sld:Title/>
-      <sld:FeatureTypeStyle>
-        <sld:Name>group 0</sld:Name>
-        <sld:FeatureTypeName>Feature</sld:FeatureTypeName>
-        <sld:SemanticTypeIdentifier>generic:geometry</sld:SemanticTypeIdentifier>
-        <sld:SemanticTypeIdentifier>simple</sld:SemanticTypeIdentifier>
-        
-        <sld:Rule>
-          <sld:Name>park</sld:Name>
-          <ogc:Filter>
-            <ogc:PropertyIsEqualTo>
-              <ogc:PropertyName>MTFCC</ogc:PropertyName>
-              <ogc:Literal>K2180</ogc:Literal>
-            </ogc:PropertyIsEqualTo>
-          </ogc:Filter>
-          <sld:MaxScaleDenominator>500000.0</sld:MaxScaleDenominator>
-          <sld:PolygonSymbolizer>
-            <sld:Fill>
-              <sld:GraphicFill>
-                <sld:Graphic>
-                  <sld:ExternalGraphic>
-                    <sld:OnlineResource
-                    xlink:type="simple"
-                    xlink:href="./img/landmarks/area/forest.png" />
-                    <sld:Format>image/png</sld:Format>
-                  </sld:ExternalGraphic>
-                </sld:Graphic>
-              </sld:GraphicFill>
-            </sld:Fill>
-          </sld:PolygonSymbolizer>
-          <VendorOption name="inclusion">mapOnly</VendorOption>
-        </sld:Rule>
-        
-        <sld:Rule>
-          <sld:Name>nationalpark</sld:Name>
-          <ogc:Filter>
-            <ogc:PropertyIsEqualTo>
-              <ogc:PropertyName>MTFCC</ogc:PropertyName>
-              <ogc:Literal>K2181</ogc:Literal>
-            </ogc:PropertyIsEqualTo>
-          </ogc:Filter>
-          <sld:MaxScaleDenominator>500000.0</sld:MaxScaleDenominator>
-          <sld:PolygonSymbolizer>
-            <sld:Fill>
-              <sld:GraphicFill>
-                <Graphic>
-                  <Mark>
-                    <WellKnownName>square</WellKnownName>
-                    <Fill>
-                      <CssParameter name="fill">#ffff00</CssParameter>
-                    </Fill>
-                  </Mark>
-                  <Size>6</Size>
-                </Graphic>
-              </sld:GraphicFill>
-            </sld:Fill>
-          </sld:PolygonSymbolizer>
-        </sld:Rule>
-      </sld:FeatureTypeStyle>
-    </sld:UserStyle>
-  </sld:UserLayer>
-</sld:StyledLayerDescriptor>
